+++ src/app/(normal)/pages/Channel.tsx
