+++ src/app/(normal)/profile/pages/DetailPage.tsx
@@ -2,7 +2,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { notFound } from 'next/navigation.js';
 
-import { ProfilePage } from '@/app/(normal)/profile/pages/Profile.js';
+import { ProfilePage } from '@/app/(normal)/pages/Profile.js';
 import { Loading } from '@/components/Loading.js';
 import type { SourceInURL } from '@/constants/enum.js';
 import { resolveSocialPlatform } from '@/helpers/resolveSocialPlatform.js';
